Replace MainActivity with CustomActivity
