+++ example/app/src/main/AndroidManifest.xml
@@ -18,7 +18,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".CustomActionActivity"></activity>
     </application>
 
 </manifest>
